feat(formarc): implement Formarc library by adding some new logic/functions
- implemented custom request func to send request with different methods(post, patch..) based on type of form(formarc)
- added new util functions to play and modify form values accordingly
- separated some logic to custom usePodarc hook
- updated naming of files and components to have correct ones
- added ignoreFor property for input fields not to send some field values in some requests, especially needed for patch requests(done partially - need to be checked for nested formValue
